reimplemented scala api

+++ scala-api/src/main/scala/com/avast/metrics/scala/api/Counting.scala
@@ -1,5 +1,5 @@
 package com.avast.metrics.scala.api
 
 trait Counting {
-  def count: Int
+  def count: Long
 }
